move action package to goordinator/action
The action package is only used by code in the goordinator, therefore make it a
subpackage.

internal/goordinator will be the package that implements the (currently only)
functionality of goordinator to receive events and run them if a jq-filter
matches.
